Move OptionalFeatureStartupListener to core module
